update processRainfall.py for kubeless
