+++ test.py
@@ -1,0 +1,3 @@
+def hello(event, context):
+    print(event)
+    return event['data']
