+++ tests/test_unit.py
@@ -1,7 +1,9 @@
 import json
 import pytest
-from handleAPIResponse import handleResponse
-import rabbitmqConnection
+from processRainfall import handleResponse
+from processRainfall import create
+from processRainfall import publish
+import processRainfall
 from unittest.mock import Mock, patch
 from tests.mocks import pika
 
@@ -11,9 +13,9 @@ from tests.mocks import pika
 def test_channel_sets_parameters(monkeypatch):
     mocked_pika = Mock()
     mocked_pika.URLParameters.return_value = 'testChannel'
-    monkeypatch.setattr('rabbitmqConnection.pika', mocked_pika)
+    monkeypatch.setattr('processRainfall.pika', mocked_pika)
 
-    rabbitmqConnection.create('testChannel host', 1)
+    create('testChannel host', 1)
 
     mocked_pika.URLParameters.assert_called_once_with('testChannel host')
 
@@ -23,9 +25,9 @@ def test_create_calls_connection(monkeypatch):
     mocked_pika = Mock()
     mocked_pika.URLParameters.return_value = 'testChannel'
     mocked_pika.BlockingConnection.return_value = pika.Connection()
-    monkeypatch.setattr('rabbitmqConnection.pika', mocked_pika)
+    monkeypatch.setattr('processRainfall.pika', mocked_pika)
 
-    rabbitmqConnection.create('testChannel host', 1)
+    create('testChannel host', 1)
 
     mocked_pika.BlockingConnection.assert_called_once_with('testChannel')
 
@@ -38,11 +40,11 @@ def test_exception_thrown_when_rabbit_down(pika_mock):
     pika_mock.BlockingConnection.side_effect = pika.exceptions.AMQPConnectionError
 
     with pytest.raises(Exception):
-        rabbitmqConnection.create('testHost', 1)
+        create('testHost', 1)
 
 @pytest.mark.unit
-@patch('handleAPIResponse.rabbitmqConnection.create', autospec=True)
-@patch('handleAPIResponse.rabbitmqConnection.publish', autospec=True)
+@patch('processRainfall.create', autospec=True)
+@patch('processRainfall.publish', autospec=True)
 def test_correct_queue_used_on_non_200_response(mock_publish, mock_create_connection):
     # mock connection to rabbitMQ and API response
     connection = pika.Connection()
@@ -59,12 +61,12 @@ def test_correct_queue_used_on_non_200_response(mock_publish, mock_create_connec
     handleResponse(response, 1)
 
     # check publish function called with correct parameters to direct message to debug queue
-    rabbitmqConnection.publish.assert_called_once_with(connection,  message, 'debug.rainfall', 'rabt-debug-exchange')
+    processRainfall.publish.assert_called_once_with(connection,  message, 'debug.rainfall', 'rabt-debug-exchange')
 
 @pytest.mark.unit
-@patch('handleAPIResponse.rabbitmqConnection.create', autospec=True)
-@patch('handleAPIResponse.rabbitmqConnection.publish', autospec=True)
-@patch('handleAPIResponse.createRainfallMessage', autospec=True)
+@patch('processRainfall.create', autospec=True)
+@patch('processRainfall.publish', autospec=True)
+@patch('processRainfall.createRainfallMessage', autospec=True)
 def test_correct_queue_used_on_200_response(mock_create_message, mock_publish, mock_create_connection):
     # mock connection to rabbitMQ, API response and message creation
     connection = pika.Connection()
@@ -80,7 +82,7 @@ def test_correct_queue_used_on_200_response(mock_create_message, mock_publish, m
     handleResponse(response, 1)
 
     # check publish function called with correct parameters to direct message to rainfall queue
-    rabbitmqConnection.publish.assert_called_once_with(connection,  message, 'rabt-rainfall-queue', 'rabt-rainfall-exchange')
+    processRainfall.publish.assert_called_once_with(connection,  message, 'rabt-rainfall-queue', 'rabt-rainfall-exchange')
 
 # https://docs.pytest.org/en/stable/pythonpath.html 
 # https://medium.com/@odelucca/recommendation-algorithm-using-python-and-rabbitmq-part-2-connecting-with-rabbitmq-aa0ec933e195
